**chore(utilities): scaffold converters modules (scalar & collection)**
Introduce empty modules to prepare splitting universal conversion logic out of `core_util`:

* `quicken_helper/utilities/converters_scalar.py` (placeholder)
* `quicken_helper/utilities/converters_collection.py` (placeholder)

No functional changes in this commit; files are scaffolding for upcoming refactor on the `feature/migrate-parsing-to-data-model` branch.
